wirequery/manager-server/src/main/kotlin/com/wirequery/manager/application/security/UserDetailsServiceImpl.kt
package com.wirequery.manager.application.security

import com.wirequery.manager.domain.user.UserService
import org.springframework.security.core.userdetails.UserDetails
import org.springframework.security.core.userdetails.UserDetailsService
import org.springframework.security.core.userdetails.UsernameNotFoundException
import org.springframework.stereotype.Service
import org.springframework.transaction.annotation.Transactional

@Service
@Transactional
class UserDetailsServiceImpl(
    private val userService: UserService,
) : UserDetailsService {
    override fun loadUserByUsername(username: String): UserDetails {
        return userService.findUserDetailsByUsername(username)
            ?: throw UsernameNotFoundException("Username not found: $username")
    }
}
